test/requirements.txt
anyio==4.8.0
APScheduler==3.11.0
certifi==2025.1.31
colorama==0.4.6
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
nest-asyncio==1.6.0
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.10
pytest==8.3.4
python-telegram-bot==21.10
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
